ui: use `u16` for max toolbar widget slots.
This lets us avoid some chances of numeric overflow, and we certainly
won't want to try to actually display more than that many slots.

+++ all-is-cubes-ui/src/ui_content/hud.rs
@@ -18,7 +18,7 @@ use crate::ui_content::{CueNotifier, VuiMessage, VuiPageState};
 use crate::vui::widgets::{self, TooltipState, WidgetBlocks};
 use crate::vui::{self, LayoutTree, UiBlocks, Widget, WidgetTree};
 
-pub(crate) const TOOLBAR_POSITIONS: usize = 10;
+pub(crate) const TOOLBAR_POSITIONS: u16 = 10;
 
 /// Ad-hoc bundle of elements needed to construct HUD UI widgets.
 ///
